moved to javax.validation for validating.
All default values for config is now in the data classes with v1
schema version. If we change the schema later new defautl values will
have to be in the new incomming data classes.

The data classes used in the template are there in order to limit the
number of places change in version schema will have to be handeled.