feature/feature context (#399)
FeatureContext introduced

This commit will introduce a new part of the feature flow called feature context. When making a feature it is possible to create a Map that is a context that is sent to all the other steps in that feature. This makes it possible to send data between the steps in a feature. 

When creating a feature a validationContext flag is sent in to tell if the context should be used only for shallow validation. When this is done you should never perform external integrations. 

Co-authored-by: Kristoffer Moberg Christensen <kristoffermoberg.christensen@skatteetaten.no>